+++ src/Instance/RedisQueue.php
@@ -3,21 +3,22 @@
  * Storage implementation for the Redis Object Cache Drop-In.
  */
 
-namespace Geniem\Queue\Storage;
+namespace Geniem\Queue\Instance;
 
 use Exception;
-use Geniem\Queue\Interfaces\EntryFetcherInterface;
+use Geniem\Queue\Interfaces\FetchableInterface;
 use Geniem\Queue\Logger;
 use Psr\Log\LoggerInterface;
-use Geniem\Queue\Interfaces\EntryHandlerInterface;
+use Geniem\Queue\Interfaces\HandleableInterface;
 use Geniem\Queue\Interfaces\EntryInterface;
+use Redis;
 
 /**
  * Class RedisCache
  *
  * @package Geniem\Queue
  */
-class RedisCache extends Base {
+class RedisQueue extends Base {
 
     /**
      * This hook name is used to add a WP Cron event for a queue.
@@ -53,7 +54,7 @@ class RedisCache extends Base {
     /**
      * Queues will be stored through the Redis instance.
      *
-     * @var \Redis
+     * @var Redis
      */
     protected $redis;
 
@@ -71,11 +72,11 @@ class RedisCache extends Base {
     /**
      * Queue constructor.
      *
-     * @param string                $name    A unique name for the queue.
-     * @param EntryFetcherInterface $fetcher The entry fetcher instance.
-     * @param EntryHandlerInterface $handler The entry handler instance.
+     * @param string              $name    A unique name for the queue.
+     * @param FetchableInterface  $fetcher The entry fetcher instance.
+     * @param HandleableInterface $handler The entry handler instance.
      */
-    public function __construct( string $name, EntryFetcherInterface $fetcher, EntryHandlerInterface $handler ) {
+    public function __construct( string $name, FetchableInterface $fetcher, HandleableInterface $handler ) {
         $this->name          = $name;
         $this->entry_fetcher = $fetcher;
         $this->entry_handler = $handler;
@@ -313,7 +314,7 @@ class RedisCache extends Base {
             $this->load_entry_handler();
 
             // Do nothing if the handler is not the correct type.
-            if ( ! $this->entry_handler instanceof EntryHandlerInterface ) {
+            if ( ! $this->entry_handler instanceof HandleableInterface ) {
                 throw new Exception( 'RedisCacheQueue - The entry handler is the wrong type.' );
             }