+++ src/Interfaces/HandleableInterface.php
@@ -5,19 +5,17 @@
 
 namespace Geniem\Queue\Interfaces;
 
-use Geniem\Queue\Entry;
-
 /**
  * Use this interface to define an entry handler.
  */
-interface EntryHandlerInterface {
+interface HandleableInterface {
 
     /**
      * The method must return an array of entries or null.
      *
-     * @param Entry $entry An entry instance.
+     * @param EntryInterface $entry An entry instance.
      * @return void
      */
-    public function handle( Entry $entry );
+    public function handle( EntryInterface $entry );
 
 }